src/requirements.txt
sentence_transformers==2.2.2
requests==2.31.0
faiss-cpu==1.7.4
colorlog==6.7.0
bs4==0.0.1